+++ src/cryptoadvance/specterext/devhelp/controller.py
@@ -6,6 +6,7 @@ from flask_login import login_required, current_user
 from cryptoadvance.specter.services.controller import user_secret_decrypted_required
 from .service import DevhelpService
 from cryptoadvance.specter.wallet import Wallet
+from cryptoadvance.specter.util.common import robust_json_dumps
 
 
 logger = logging.getLogger(__name__)
@@ -68,3 +69,47 @@ def settings_post():
         wallet = current_user.wallet_manager.get_by_alias(used_wallet_alias)
         DevhelpService.set_associated_wallet(wallet)
     return redirect(url_for(f"{DevhelpService.get_blueprint_name()}.settings_get"))
+
+
+NEVER_CALLED_PYTHON_COMMAND = True
+
+
+@devhelp_endpoint.route("/python_command", methods=["POST"])
+@login_required
+def python_command():
+    global NEVER_CALLED_PYTHON_COMMAND
+    if NEVER_CALLED_PYTHON_COMMAND:
+        NEVER_CALLED_PYTHON_COMMAND = False
+        return robust_json_dumps(
+            "!!!DANGER!!!\n"
+            "This command allows arbitrary access to Specter.\n"
+            "You can irreparabily damage your specter configuration and\n"
+            "all funds on !!!HOT!!! wallets can be lost!\n"
+            "Please use it with extreme care!!! Run your command again to execute it.\n\n"
+            "--> Never copy&paste anything you do not FULLY understand. <--"
+        )
+    if current_user != "admin" or not app.specter.user_manager.user.is_admin:
+        return robust_json_dumps(f"Access forbidden for user '{current_user}'!")
+    if not app.config["DEVELOPER_JAVASCRIPT_PYTHON_CONSOLE"]:
+        return robust_json_dumps(
+            "DEVELOPER_JAVASCRIPT_PYTHON_CONSOLE disabled in Specter configuration.  "
+            "This is an advanced option and should be used with great care!!!"
+        )
+    if request.method != "POST":
+        return robust_json_dumps("Not a 'POST' command.")
+
+    # The following commented lines are a further restriction of this endpoint, by limiting it only to regtest and testnet
+    # uncomment these lines to enable the restriction
+    # ----------------------------------------------
+    # allowed_chains = ['regtest', 'testnet', 'liquidtestnet', 'liquidregtest']
+    # if app.specter.chain not in allowed_chains:
+    #     return robust_json_dumps(f"This command is only allowed for {allowed_chains}. "
+    #                              "The current chain is {app.specter.chain}")
+    # ----------------------------------------------
+
+    command = request.form["command"]
+    result = app.console.exec_command(command)
+    try:
+        return robust_json_dumps(result)
+    except:
+        return robust_json_dumps(str(result))
